Prototype rust code
Initial cut of the code in a fresh repository to prevent any overlap
or concern about mixing with the current open source breathe code.

This code has been written from scratch.

It works in a sphinx set up but I have not copied in that code to here
and will set it up from scratch.