Add basic SectionDef and remove 'ClassInfo'
I was hoping to have an abstraction layer between the backend (doxygen,
clang-doc, etc) and frontend but it seems like it is too much hassle
to create an independent representation of everything that could be
rendered right now. Instead we switch to directly rendering the structs
that we've parsed out of the doxygen xml. We might be able to make
better judgements about clang-doc if we ever move to that.