Add new example for class methods
I want to better understand the xml for method parameters without
ruining the nutshell example. So we create an example explicitly for
class methods.

Also switch to using 'Doxyfile' as the name of the doxygen config file
as that is the default that doxygen looks for.

Also support explicitly naming the project in the rst so we can have
different directives focusing on different projects.

Also change directive name to 'doxygenclass' to provide a better
migration from the old breathe project.