Add support for section def and member def references
We generate the target nodes and the references nodes and sphinx hooks
it all up.

We have passed the document through to the node builder (after some
refactoring) so that we can call 'note_explicit_target' but we're
not actually calling it and the references are working at the moment
so I'm not sure what it is for. Presumably required for some more
complex example.