Create script to generate doxygen compound types
A very hacky script which generates the elements.rs from the compound
xsd file.

This does not build with the rest of the code and needs to be adjusted
in various ways but the output is valid in some kind of way so it is
interesting to track the progress from here.