Make generate script "comment aware"
It parses the rust code (with a hacky regex approach) to figure out
which struct fields have been commented out and then preserves those
comments in the regenerated file so that we don't have to worry about
restoring comments after generation.

We have to change the field names back to the auto-created ones so
that the lookups work properly. Field name edits should be done via
the script to preserve this workflow.