Set up maturin as the build system
Using the 'src' and 'rust' strategy to separate the python and rust
code and also not have the python code at the top level (to avoid
import confusion) and also have the rust code build as a submodule of
the python package (breathe_plus.backend in this case).

This is described as a 'mixed' python and rust project in the maturin
documentation.