Add struct & function directives with examples
The struct code can be mostly unified with the class handling but
the functions come as members of "file" level compounds so we have
to look for the right compound and member id in the index file and
then find the member again in the compound.xml file.

Also stop linking the documentation directly to the breathe module
using 'sys.path' alterations. Instead we rely on 'maturin develop' to
make the breathe_plus package available in the python virtual env
and assume it'll be picked up from there.

Also generate the index/elements.rs file using the generate.py script
and change the xml parsing to something a little smarter given the
experience and helpers we have from the compound parsing code.