Switch to 'tracing' crate
And use it the same way as the logging crate. I think tracing is most
useful for multi-threaded situations and we're not necessarily there
but it is useful to have in place in case we do that and in case we
need to offer the more features like structured logging and things
which I believe are better supported by tracing.