Respect the use="optional" xsd entries
It seems that it can be used to indicate that an attribute is not
required. It might even be the default state but for the moment we
don't treat it that way.

I guess this is the attribute equivalent of the 'min bounds'
'max bounds' values for child nodes but you can never have more than
one of the same attribute so it is just optional or not.