Add basic build.rs file
Aimed at eventually being a from-xsd rust module generation. For the
moment it just dumps a ast generated from a sample of rust code using
the 'quote' crate. We might not need the quote crate in the end, I
don't have a complete picture of how all this should work.