Attempt to parse groups correctly
Including groups within groups.

Manually we'd done this is in awkward way but here if the element is
not a direct member of the group then we fall back using match logic
to trying the sub-group. This would fail if there was more than one
sub-group but there isn't in our xsd.