Handle empty tags
We can't do a full parse on empty tags as the next event is not
something that is useful within the context of the tag itself as
the tag is already 'over'.

So we introduce a parse_empty helper and call that when relevant so
that we can just look for attributes on the tag and nothing else.

Also use the try_get_attribute helper method as that is doing what
we're doing already.