Add more error logging for unhandled data
So that we get a full idea of what isn't being handled when we run
the build of some docs and we can work through the messages to make
sure everything is handled.

Also deferring to render_doc_cmd_group for all doc_cmd_group data
instead of intercepting some of it.