Improve parameter description rendering
We render it and then extract the nodes from inside the Paragraph as
we want the content inline instead of as new block.

The old Breathe code does something similar.

I've removed 'mixed=True' from the descriptionType in the xsd as it
seems like the contents could reasonable be parsed into something that
doesn't required text in there as all text should be captured inside
the 'para' tags but we'll see. We might need to change it back.