Rename 'documentation' to 'examples/sphinx'
This is to acknowledge that it isn't really our documentation, it is
just a bunch of samples. We merge them together in the Breathe project
but that is a bit ugly so I'd like to try another approach here and
leave room for clear and specific documentation and then a separate
location for examples/testing.