Basic support for content-only and inner options for groups
As supported by Breathe.

The inner support is pretty hacky and should be replaced with a 'filter'
concept and system that can help us out with other things.

The content-only might also be done with filters but I'm not sure.
Maybe more the 'matchers' concept from the previous code base.

Introduced the XmlLoader' concept which wraps the Cache which meant a
lot of life time troubles so I've switch the cache out to be an
Arc-Mutex arrangement which seems to work fine.