Add rustfmt to github workflow
And run it.

We need it installed to allow the actual build to happen as the build.rs
script attempts to format the resulting code with rustfmt. Perhaps we
should make that optional but it is ok for the moment.

We also take the opportunity to run 'cargo fmt --check' on the code
base.