Introduce domain concept and defer to Sphinx domains
I've been trying to achieve it in other ways but it seems hard to avoid
using Sphinx domain code to play nicely with Sphinx domains. In
particular, Sphinx role like :cpp:func:`` generate a custom reference
string which it expects to target and which is going to be hard to
replicate without using the domain Python code.

So we change so that under certain situations if we have a valid domain
then we render the compound's contents but defer the 'title' parts of
the compound to the domain rendering. We then add the contents to the
render domain output and use that as the final result.

It is modelled on the approach in Breathe and uses some of the code
which I've moved into the 'copied' module which I'll update to be very
explicit about the license for that module.

As the domain rendering produces a node list we have to change the
python code to thing in terms of lists rather than individual nodes.
The current approach with the tuple is a bit ugly but it works.

To do the domain rendering we end up "running" a docutils directive
from the sphinx domain code and in order to make that work we have to
thread the arguments from our own directive invocations down through
to the domain directive invocation so that it has valid values for
everything. This is a bit ugly but seems necessary.