Insert target into domain output
So that we can still link to the output from other doxygen generated
parts as they only have access to the doxygen refid and not the domain
generated ref.

Also use the concatenated directive name as noted in the breathe code
as otherwise the domain output doesn't include some of the references
to the domain name that we would expect. ie. it seems to pull data out
of the directive name rather than rely on how it is being called.