Support rendering parameter lists as field lists
Like breathe does.

Involves making sure that we categorize the parameter list render
output so that we can extract further up the stack when needed.

A bit of faff but it allows us to put all special nodes like parameter
list and probably 'returns' into a field list altogether instead of
having a field list per item. Though we could probably work to have
only neighbouring items in the same list one day.

Ended up fixing the 'replace children on parent' logic in domains.py.
We weren't using an approach that would allow docutils to properly
set up the children so none of the children knew there parents which
caused issues with the field lists (but not with previous nodes.)