Track list nested and change list numbering type
As we descend down the lists that we encounter, we increment a tracking
value on the context and use that to calculate the list numbering type.

We also allow for the xml informing the choice but I don't have an
example of that working so I'm not sure how that comes about.