Add support for domains by file extension
It seems that doxygen doesn't label files as being in the C language
and so if we want to target the C domain in Sphinx then we have to
take extra steps to indicate as much to the build so we implement
the 'domain by extension' concept and allow us to label certain files
as C.

We then move all our example files to be C++ as they have things like
'classes' in them which need C++ support but we'll introduce C files
soon.

Basing the 'domain by extension' concept off the doxygen xml 'location'
nodes forced us to notice that we weren't parsing them correctly so
we have to extend the xsd handling stuff with some ugly copy and paste
code which handles the empty node instances for non-mixed-content
'struct' based data types.