Basic support for defines/macros
We're not properly rendering macros with arguments here but that'll be
the next step.

We have to introduce a new string to the domains.py for each type as
doxygen thinks of these things as 'defines' and sphinx thinks of them
as 'macros' so we map from one to the other.