Remove special anonymous union code
As there doesn't seem to be enough information in the doxygen output
to put properly identify the unions and their members. I hope to return
to this and have some code that tries to parse the source code at the
appropriate lines to figure it out a bit better but it seems better
to fix up some other issues first.