Add initial implementation of rebuilds checks
We want to tell Sphinx what files to rebuild if the doxygen output
has changed. We check the mtime and the hash of the doxygen files as
doxygen will happily overwrite xml files with the same content so it
is a better experience if we double check that their contents are
the same as previously.

This is only a partial implementation as it does the work for classes
(and not even inner classes) and nothing else. The main issue is that
the render has to collect the files that it interacts with and store
the relationships as data in the Sphinx env to be checked on the next
build.

Also introduces use of the logger on the Sphinx side which allows us
to add some logs and keep them around. Builds with -vv will output the
logs but they won't appear by default.