Switch strategy to 'tracked cache'
Instead of expecting the render methods to report what xml files they
have accessed we change to using a cache that'll track the accesses
to xml files and let us query them.

This reduces the intrusion into the code and means that we have our
accesses tracked for us.

We move the caches types into a separate file so that we can be
stricter about public/private accesses and we clean up the direct
accesses to the cache.inner in favour of using methods for it.

I tried to make a Cache trait for various parts of the system to
rely on but that hasn't worked out as my understanding of Boxed traits
is pretty weak still and that seemed to be necessary in some parts.

I hope we can revisit this if necessary in the future. For the moment
the XmlLoader helps to abstract away the specific cache used.