Optional support for linkcode extension
We provide a helper for using the linkcode extension so that we
resolve the links based on some information that we pack into the
'names' attributes on signature nodes in our domain calls as those
at the only ones that are supported by the linkcode extension.

I don't know if we should be treating the 'names' attribute differently
or if it might cause issues down the line but it seems to be working
in this initial form for the test docs.