Improve node calling convention names
The previous names where poorly made up and didn't reflect any deeper
knowledge of the system.

After looking more closely at the docutils nodes, the different is
whether the node inherits straight from Element or via TextElement so
we name the different strategies after those.

Then we include 'Function' as we're just calling the end result
directly in that case.

We also examine the nodes and change some of the calling conventions
appropriately. Mostly Element to TextElement where needed.