Include mermaid image in doctree
We create an 'image' node with the generated image referenced in it
and we make sure that we write the image to a reasonable directory (ie.
a sibling directory of the doctrees folder.) I can't see a direct way
to get the 'build' directory itself but going via the doctrees folder
seems ok.

The image is currently too large and unwieldy in the output. We need
to figure out ways to tame it a bit or push it off to a separate page
or into appendicies or something. Or collapse it somehow for html and
leave it out for PDF. Something anyway.