Add initial version of doxygen-inheritance-graph directive
And a config option for the mermaid executable execution components.

I've realised that we're allowed '-'s in directive names and given
that 'doxygeninheritancegraph' is not that readable it seems worth
using them which means it is worth using them for the other directives
too so we leave the old names as fallbacks but add dashed versions as
new directives.

We also the directive to render doxygen inheritance graphs as they
are a bit too bulky to throw into the class output directly and this
gives us some flexibility for the moment.

The image output is large and uncontrolled and needs to be improved.

During development I was getting an unhelpful file system error which
turned out to be because the code was trying to run 'mmdc' and that
wasn't on the path. We change the error message to be better and allow
users to override the "execution components" for the mmdc command so
we can (as shown here) use 'npx mmdc' if we want and do it the proper
way with a list of strings instead of using a single string.