Generate third party licenses file
And name it such that it is included in the maturin built wheel for
distribution.

We could generate this on the fly during the release process but for
the moment I want it checked in.

Bit of a nightmare as some licenses have windows line endings which
'cargo bundle-licenses' preserves but then Python strips out when
writing or reading the file which meant that the 'check' mode would
repeatedly fail. Added a 'replace' call to strip them out and it seems
to work now.