Add 'objtype' and 'domain' attributes to 'desc' node
As pointed out in issue #3, 'desc' nodes are expected to have 'objtype'
and 'domain' attributes as indicated by this comment:

  https://github.com/sphinx-doc/sphinx/blob/d3c91f951255c6729a53e38c895ddc0af036b5b9/sphinx/addnodes.py#L143-L146

Here we change our definition to Desc to require them though only
require an Option<Domain> so that we control the fallback site.

We make sure they are present for all the 'desc' nodes that we directly
construct and I assume that the sphinx domain directives add them
correctly.